+++ framework/Furion/DatabaseAccessor/Repositories/MSRepository.cs
@@ -1,7 +1,124 @@
 ﻿using Furion.DependencyInjection;
+using System;
+using System.Security.Cryptography;
 
 namespace Furion.DatabaseAccessor
 {
+    /// <summary>
+    /// 默认主库主从仓储
+    /// </summary>
+    [SkipScan]
+    public partial class MSRepository : MSRepository<MasterDbContextLocator>, IMSRepository
+    {
+        /// <summary>
+        /// 构造函数
+        /// </summary>
+        /// <param name="serviceProvider"></param>
+        /// <param name="repository">非泛型仓储</param>
+        public MSRepository(IServiceProvider serviceProvider
+            , IRepository repository) : base(serviceProvider, repository)
+        {
+        }
+    }
+
+    /// <summary>
+    /// 主从库仓储
+    /// </summary>
+    /// <typeparam name="TMasterDbContextLocator">主库</typeparam>
+    [SkipScan]
+    public partial class MSRepository<TMasterDbContextLocator> : IMSRepository<TMasterDbContextLocator>
+        where TMasterDbContextLocator : class, IDbContextLocator
+    {
+        /// <summary>
+        /// 服务提供器
+        /// </summary>
+        private readonly IServiceProvider _serviceProvider;
+
+        /// <summary>
+        /// 非泛型仓储
+        /// </summary>
+        private readonly IRepository _repository;
+
+        /// <summary>
+        /// 构造函数
+        /// </summary>
+        /// <param name="serviceProvider"></param>
+        /// <param name="repository">非泛型仓储</param>
+        public MSRepository(IServiceProvider serviceProvider
+            , IRepository repository)
+        {
+            _serviceProvider = serviceProvider;
+            _repository = repository;
+        }
+
+        /// <summary>
+        /// 获取主库仓储
+        /// </summary>
+        /// <typeparam name="TEntity">实体类型</typeparam>
+        /// <returns></returns>
+        public virtual IRepository<TEntity, TMasterDbContextLocator> Master<TEntity>()
+            where TEntity : class, IPrivateEntity, new()
+        {
+            return _repository.Change<TEntity, TMasterDbContextLocator>();
+        }
+
+        /// <summary>
+        /// 动态获取从库（随机）
+        /// </summary>
+        /// <typeparam name="TEntity">实体类型</typeparam>
+        /// <returns></returns>
+        public virtual IReadableRepository<TEntity> Slave<TEntity>()
+            where TEntity : class, IPrivateEntity, new()
+        {
+            // 判断数据库主库是否注册
+            var isRegister = Penetrates.DbContextWithLocatorCached.TryGetValue(typeof(TMasterDbContextLocator), out Type dbContextType);
+            if (!isRegister) throw new InvalidCastException($" The locator `{typeof(TMasterDbContextLocator).Name}` is not bind.");
+
+            // 获取主库贴的特性
+            var appDbContextAttribute = DbProvider.GetAppDbContextAttribute(dbContextType);
+
+            // 获取从库列表
+            var slaveDbContextLocators = appDbContextAttribute.SlaveDbContextLocators;
+
+            // 如果没有定义从库定位器，则抛出异常
+            if (slaveDbContextLocators == null || slaveDbContextLocators.Length == 0) throw new InvalidOperationException("Not found slave locators.");
+
+            // 如果只配置了一个从库，直接返回
+            if (slaveDbContextLocators.Length == 1) return Slave<TEntity>(() => slaveDbContextLocators[0]);
+
+            // 获取随机从库索引
+            var index = RandomNumberGenerator.GetInt32(0, slaveDbContextLocators.Length - 1);
+
+            // 返回随机从库
+            return Slave<TEntity>(() => slaveDbContextLocators[index]); ;
+        }
+
+        /// <summary>
+        /// 动态获取从库（自定义）
+        /// </summary>
+        /// <typeparam name="TEntity">实体类型</typeparam>
+        /// <returns></returns>
+        public virtual IReadableRepository<TEntity> Slave<TEntity>(Func<Type> locatorHandle)
+            where TEntity : class, IPrivateEntity, new()
+        {
+            if (locatorHandle == null) throw new ArgumentNullException(nameof(locatorHandle));
+
+            // 获取定位器类型
+            var dbContextLocatorType = locatorHandle();
+            if (!typeof(IDbContextLocator).IsAssignableFrom(dbContextLocatorType)) throw new InvalidCastException($"{dbContextLocatorType.Name} is not assignable from {nameof(IDbContextLocator)}.");
+
+            // 判断从库定位器是否绑定
+            var isRegister = Penetrates.DbContextWithLocatorCached.TryGetValue(dbContextLocatorType, out _);
+            if (!isRegister) throw new InvalidCastException($" The slave locator `{dbContextLocatorType.Name}` is not bind.");
+
+            // 解析从库定位器
+            var repository = _serviceProvider.GetService(typeof(IRepository<,>).MakeGenericType(typeof(TEntity), dbContextLocatorType)) as IPrivateRepository<TEntity>;
+
+            // 返回从库仓储
+            return repository.Constraint<IReadableRepository<TEntity>>();
+        }
+    }
+
     /// <summary>
     /// 主从库仓储
     /// </summary>